rename tag_parser and accent_parser
